% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynlib.R
\name{classifySpatiallyByBlocks}
\alias{classifySpatiallyByBlocks}
\title{classifySpatiallyByBlocks}
\usage{
classifySpatiallyByBlocks(df, lon, lat, size)
}
\arguments{
\item{size}{}
}
\description{
This is also a Test
}
