% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynlib.R
\name{rasterMakerAdjust}
\alias{rasterMakerAdjust}
\title{rasterMakerAdjust}
\usage{
rasterMakerAdjust(df, minX, maxY, nCols, nRows, sizeOfGen, sizeOfList,
  degree, block, horizontal, vertical)
}
\arguments{
\item{vertical}{}
}
\description{
This is a Test
}
