\name{rspatial-package}
\alias{rspatial-package}

\alias{rspatial}

\docType{package}

\title{rspatial}

\description{
This package provides simplified data access for examples on rspatial.org}

\author{
Robert Hijmans and Alex Mandel
Maintainer: Robert J. Hijmans <r.hijmans@gmail.com>
}

\references{
https://rspatial.org
}

\keyword{ package }
\keyword{ spatial }

